Attempted fix of missing rise or set times.
It was possible not to find a rise or set time if the starting
point was close to transit (within the range of setting HA).
Implemented a new scheme where instead of taking meridians to
either side of the starting point, we first take the one to the
desired side.  We then step again in the same direction unless
it was the wrong way, in which case we go the other way.