Change strptime format for 3-letter month
The Astro::Coords::Elements constructor documentation indicates
that EPOCH can be given as a string with the month as a 3-letter
string (e.g. "Apr").  This seems to require a format specifier of
"%b" rather than "%B".  (The string is parsed with Time::Piece's own
BSD-style strptime function which treats "%b" and "%B" differently.)