Update t/angle.t test script
This changes a test of a numeric value (45 degrees) from using "is"
to using "delta_ok", due to recent test failures reported when the
string form isn't quite "45", e.g. "44.9999999999999982".

Also set "str_ndp" before checking the default stringification (so
that the test doesn't depend on the default NDP not changing) and give
the negation test a name.