Angle: use str_ndp for in_format('array')
This is so that the extra precision (3 d.p. by default) is used
when requesting RA in array format.

Note: unsure if this should use the class NDP or instance 'str_ndp'
since the documentation for 'str_ndp' mentions "stringifying" -- does
that mean just 'string' or also in_format('array')?

Also note: the documentation for NDP says it should apply to both
the 'string' and 'components' methods.  However the documentation
for the 'components' method says it should ignore 'str_ndp'.
While confusing, this doesn't directly relate to this commit,
which applies to how 'in_format' calls 'components'.