Correct sense of radio rv check for redshift calc.
The documentation for the "redshift" method says that the calculation
will only be performed for radio velocities if the value is small.
However the sense of the test appears to have been flipped in
commit 0f349594e1 so that the calculation would instead only
be performed for large velocities.  Restore the original sense
of the check and add an example to the test suite to ensure
the method is operating for small velocities.